requirements.txt
openai
anthropic
ffmpeg-python
imageio_ffmpeg